# rest/swaggerui/swagger.yaml
swagger: "2.0"
info:
  version: 0.1.0
  title: Openstack Big Data tools
  description: "Данная система предоставляет REST API для создания кластеров с заданным списком зарегестрированных сервисов в облачной среде OpenStack."
  license:
    name: License Here
  #    url:
  contact:
    email: modis@ispras.ru
host: host.here.com
basePath: /home
schemes:
  - http
  - https
...
paths:
...
  /projects/{projectId}/clusters/{clusterName}:
    get:
      tags:
        - cluster
      summary: Получение информации о конкретном кластере
      description: "Метод возращает список кластер с именем clusterName."
      operationId: ClustersGet
      parameters:
        - name: clusterName
          description: "Имя кластера."
          in: path
          type: string
          required: true
        - name: projectId
          description: "ID проекта."
          in: path
          type: string
          required: true
      produces:
        - application/json
      responses:
        200:
          description: "OK"
          schema:
            $ref: '#/definitions/Cluster'
        204:
          description: "Нет результатов, соответствующих запросу."
        400:
          description: "Некорректный запрос, проверьте параметры"
        401:
          description: "Не пройдена авторизация"
        403:
          description: "Не достаточно прав"
        404:
          description: "Not found. (может заменять 401, 403)"
    put:
      tags:
        - cluster
      summary: Обновление конфигурации существующего кластера
      operationId: updateCluster
      parameters:
        - name: clusterName
          in: path
          type: string
          required: true
        - name: configuration
          in: body
          schema:
            $ref: '#/definitions/Cluster'
        - name: projectId
          description: "ID проекта."
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
    delete:
      tags:
        - cluster
      summary: Удаленние существующего кластера
      operationId: deleteCluster
      parameters:
        - name: clusterName
          type: string
          in: path
          required: true
        - name: projectId
          description: "ID проекта."
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  /projects/{projectId}/clusters/{clusterName}/status:
    get:
      tags:
        - cluster
      summary: Получение информации о статусе конкретного кластера
      description: "Метод возращает статус кластера с именем clusterName."
      parameters:
        - name: clusterName
          description: "Имя кластера."
          in: path
          type: string
          required: true
        - name: projectId
          description: "ID проекта."
          in: path
          type: string
          required: true
      produces:
        - application/json
      responses:
        200:
          description: "OK"
...
definitions:
...
  Cluster:
    type: object
    example:
      {
        "ID":"uuid",
        "ProjectID":"projectUuid",
        "Name":"clusterName-projectName",
        "DisplayName":"name",
        "EntityStatus":"status",
        "Services":[
        {
          "ID":"uuid",
          "Name":"serviceName",
          "DisplayName":"name",
          "EntityStatus":"status",
          "Type":"serviceType",
          "Version":"version",
          "Config":{

          },
          "Description":"someDescription",
          "URL": "serviceURL"
        }
        ],
        "NHosts":3,
        "MasterIP": "masterHost",
        "Description":"someDescription"
      }
...